split runners into submodules
